Cleanup imports for WGGermline pipelines
